Ajustes para deploy
- Remove botão de redirecionamento nas telas de vendas e melhora Dashboard
- Fixa globalização como pt-BR
- Ajusta o DbContext e refaz as migrações
- Bloqueia exclusão de dados relacionados e importantes para visualização dos relatórios

+++ codigo-fonte/BrechoLaFripAtelier/Pages/Partners/Delete.cshtml.cs
@@ -47,12 +47,22 @@ namespace BrechoLaFripAtelier.Pages.Partners
 
             if (partner != null)
             {
+                if (HasAssociatedProducts(partner))
+                {
+                    return RedirectToPage("../Admins/AccessDenied");
+                }
+
                 Partner = partner;
                 _context.Partners.Remove(Partner);
                 await _context.SaveChangesAsync();
             }
 
             return RedirectToPage("./Index");
+        }
+
+        private bool HasAssociatedProducts(Partner partner)
+        {
+            return _context.Products.Any(s => s.PartnerId == partner.Id);
         }
     }
 }
